Refactor navigation code into multiple files
